COM-1404-final fix step creation
