+++ src/controllers/subProgramController.js
@@ -1,5 +1,6 @@
 const Boom = require('@hapi/boom');
 const SubProgramHelper = require('../helpers/subPrograms');
+const StepHelper = require('../helpers/steps');
 const translate = require('../helpers/translate');
 
 const { language } = translate;
@@ -15,7 +16,21 @@ const update = async (req) => {
   }
 };
 
+const addStep = async (req) => {
+  try {
+    await StepHelper.addStep(req.params._id, req.payload);
+
+    return {
+      message: translate[language].subProgramUpdated,
+    };
+  } catch (e) {
+    req.log('error', e);
+    return Boom.isBoom(e) ? e : Boom.badImplementation(e);
+  }
+};
+
 
 module.exports = {
   update,
+  addStep,
 };
