+++ src/helpers/steps.js
@@ -1,13 +1,13 @@
 const Boom = require('@hapi/boom');
 const Step = require('../models/Step');
-const Program = require('../models/Program');
+const SubProgram = require('../models/SubProgram');
 
 exports.updateStep = async (stepId, payload) => Step.updateOne({ _id: stepId }, { $set: payload });
 
-exports.addStep = async (programId, payload) => {
-  const program = await Program.countDocuments({ _id: programId });
-  if (!program) throw Boom.badRequest();
+exports.addStep = async (subProgramId, payload) => {
+  const subProgram = await SubProgram.countDocuments({ _id: subProgramId });
+  if (!subProgram) throw Boom.badRequest();
 
   const step = await Step.create(payload);
-  await Program.updateOne({ _id: programId }, { $push: { steps: step._id } });
+  await SubProgram.updateOne({ _id: subProgramId }, { $push: { steps: step._id } });
 };
