+++ src/routes/subPrograms.js
@@ -2,7 +2,8 @@
 
 const Joi = require('joi');
 Joi.objectId = require('joi-objectid')(Joi);
-const { update } = require('../controllers/subProgramController');
+const { update, addStep } = require('../controllers/subProgramController');
+const { STEP_TYPES } = require('../models/Step');
 
 exports.plugin = {
   name: 'routes-sub-programs',
@@ -18,6 +19,19 @@ exports.plugin = {
         auth: { scope: ['programs:edit'] },
       },
       handler: update,
+    });
+
+    server.route({
+      method: 'POST',
+      path: '/{_id}/steps',
+      options: {
+        validate: {
+          params: Joi.object({ _id: Joi.objectId().required() }),
+          payload: Joi.object({ name: Joi.string().required(), type: Joi.string().required().valid(...STEP_TYPES) }),
+        },
+        auth: { scope: ['programs:edit'] },
+      },
+      handler: addStep,
     });
   },
 };
